refactor: cleanup Resizer, ResizeHandle & SlideElement
